+++ trunk/ncsa/hdf/hdf5lib/HDF5Constants.java
@@ -576,7 +576,7 @@ public class HDF5Constants {
     public static final long H5T_UNIX_D64BE = H5T_UNIX_D64BE();
     public static final long H5T_UNIX_D64LE = H5T_UNIX_D64LE();
     public static final long H5T_VARIABLE = H5T_VARIABLE();
-    public static final long H5T_VLEN = H5T_VLEN();
+    public static final int H5T_VLEN = H5T_VLEN();
     public static final int H5Z_CB_CONT = H5Z_CB_CONT();
     public static final int H5Z_CB_ERROR = H5Z_CB_ERROR();
     public static final int H5Z_CB_FAIL = H5Z_CB_FAIL();
@@ -1701,7 +1701,7 @@ public class HDF5Constants {
 
     private static native final long H5T_VARIABLE();
 
-    private static native final long H5T_VLEN();
+    private static native final int H5T_VLEN();
 
     private static native final int H5Z_CB_CONT();
